@@ -10,6 +10,7 @@
     "build": "rimraf ./lib && tsc",
     "lint": "prettier --check '**/*.{js,ts,mjs,cjs}' && eslint tests/ src/",
     "fmt": "prettier --write '**/*.{js,ts,mjs,cjs}'",
+    "prepare": "husky || true",
     "test": "jest --env node",
     "release": "release-it --github.release",
     "release:ci": "npm run release -- --ci --npm.skipChecks --no-git.requireCleanWorkingDir",
@@ -27,32 +28,33 @@
     "yargs": "^17.7.2"
   },
   "devDependencies": {
-    "@eslint/js": "^9.4.0",
+    "@eslint/js": "^9.13.0",
     "@tsconfig/node22": "^22.0.0",
     "@types/chai": "^4.3.16",
+    "@types/eslint__js": "^8.42.3",
+    "@types/eslint-config-prettier": "6.11.3",
     "@types/jest": "^29.5.12",
     "@types/jquery": "^3.5.30",
     "@types/lodash": "^4.17.4",
     "@types/mock-fs": "^4.13.4",
     "@types/node": "^20.14.2",
     "@types/npm": "^7.19.3",
-    "@typescript-eslint/eslint-plugin": "^7.12.0",
-    "@typescript-eslint/parser": "^7.12.0",
-    "eslint": "^8.56.0",
+    "eslint": "^9.13.0",
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-jest": "^28.5.0",
     "eslint-plugin-prettier": "^5.1.3",
-    "husky": "^9.0.11",
+    "husky": "^9.1.6",
     "jest": "^29.7.0",
     "mock-fs": "^5.2.0",
     "npm-run-all": "^4.1.5",
-    "prettier": "^3.3.1",
+    "prettier": "3.3.1",
+    "pretty-quick": "^4.0.0",
     "release-it": "^17.3.0",
     "rimraf": "^5.0.7",
     "spawk": "^1.8.2",
     "ts-jest": "^29.1.4",
     "typescript": "^5.4.5",
-    "typescript-eslint": "^7.12.0"
+    "typescript-eslint": "^8.10.0"
   },
   "files": [
     "lib",
