@@ -1,8 +1,9 @@
-const ts = require('typescript-eslint');
-const js = require('@eslint/js');
-const prettier = require('eslint-config-prettier');
+// @ts-check
+import ts from 'typescript-eslint';
+import js from '@eslint/js';
+import prettier from 'eslint-config-prettier';
 
-module.exports = ts.config(
+export default ts.config(
   js.configs.recommended,
   ...ts.configs.recommended,
   prettier,
@@ -10,9 +11,9 @@ module.exports = ts.config(
     ignores: ['lib/**'],
   },
   {
-    files: ['*.js'],
+    files: ['**/*.ts'],
     rules: {
-      '@typescript-eslint/no-var-requires': 'off',
+      '@typescript-eslint/no-require-imports': 'off',
     },
   },
   {
